_site/requirements.txt
begins
python-slugify